Create Hamilton cycles for participants
