+++ requirements.txt
@@ -1,2 +1,4 @@
 geographiclib==1.52
 geopy==2.2.0
+numpy==1.23.4
+python-tsp==0.3.1
